Docker_Labs/Multi_stage_docker/requirements.txt
tensorflow==2.15.0
scikit-learn
keras
Flask
requests
Pillow